requirements.dev.txt
lazy-ninja
inflect>=7.5.0
openapi-generator-cli[jdk4py]>=7.0.0
pytest-cov>=6.0.0
pyjwt>=2.0
colorama>=0.4.6
pytest-django>=4.10.0
mkdocs>=1.6.1
mkdocs-autorefs>=1.4.1
mkdocs-get-deps>=0.2.0
mkdocs-material>=9.6.9
mkdocs-material-extensions>=1.3.1
mkdocs-redirects>=1.2.2
mkdocs-static-i18n>=1.3.0
mkdocstrings>=0.29.0
mkdocstrings-python>=1.16.7
